queen_tiana_woodward/requirements.txt
pyspark
Pillow
numpy